+++ Server/BankingServer/Account.DAL/Interfaces/IOperationRepository.cs
@@ -5,7 +5,7 @@ namespace Account.DAL.Interfaces
     public interface IOperationRepository
     {
         Task AddToHistoryTableAsync(OperationEntity opEntityFrom, OperationEntity opEntityTo);
-        Task<int> GetAccountBalanceByAccountIdAsync(int id);
+        Task<float> GetAccountBalanceByAccountIdAsync(int id);
         Task<int> GetOtherSideIdAsync(Guid transactionId, int accountId);
         Task<List<OperationEntity>> GetOperationsByAccountIdAsync(int id);
         Task<List<OperationEntity>> getOpertaionsByFilterPageAsync(int accountId,int pageNumber,int numOfRecrds);
