+++ Server/BankingServer/Account.DTO/OperationDto.cs
@@ -4,8 +4,8 @@
     {
         public bool DebitOrCredit { get; set; }
         public int OtherSide { get; set; }
-        public int Amount { get; set; }
-        public int Balance { get; set; }
+        public float Amount { get; set; }
+        public float Balance { get; set; }
         public DateTime OperationTime { get; set; }
 
     }
