+++ Server/BankingServer/Account.Services/Interfaces/IAccountSagaService.cs
@@ -3,7 +3,7 @@
     public interface IAccountSagaService
     {
         Task<bool> CheckIdValidAsync(int id);
-        Task<bool> CheckBalanceAsync(int id, int amount);
-        Task UpdateBalanceAsync(int fromAccount, int toAccount, int amount);
+        Task<bool> CheckBalanceAsync(int id, float amount);
+        Task UpdateBalanceAsync(int fromAccount, int toAccount, float amount);
     }
 }
