+++ Server/BankingServer/Transaction.DAL/Entities/TransactionEntity.cs
@@ -18,7 +18,7 @@ namespace Transaction.DAL.Entities
 
         [Required]
         [Range(1, 100000000)]
-        public int Amount { get; set; }
+        public float Amount { get; set; }
 
         [Required]
         public DateTime DateOfTransaction { get; set; }
